% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprint.R
\name{cprint}
\alias{cprint}
\alias{cprint.numeric}
\alias{cprint.character}
\alias{cprint.data.frame}
\title{Copy Print}
\usage{
cprint(x, ...)

\method{cprint}{numeric}(x, ...)

\method{cprint}{character}(x, ...)

\method{cprint}{data.frame}(x, ...)
}
\arguments{
\item{x}{An appropriate R object.}

\item{...}{Additional optional arguments. (Currently ignored.)}
}
\description{
\code{cprint()}  prints common R data structures in a format which can be
used as input to the interpreter.
}
\examples{

cprint(head(iris))  # output can easily be copied into a script file
}
